backend/requirements.txt
fastapi>=0.110.0
sqlmodel>=0.0.16
sqlalchemy>=2.0.29
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
pydantic-settings>=2.2.1
python-multipart>=0.0.20
python-dotenv>=1.0.1
PyJWT>=2.8.0
python-jose[cryptography]>=3.3.0
pytest>=8.2.0
httpx>=0.27.0
openai-agents>=0.0.7
mcp[cli]>=1.0.0
